Fix SoftAssertionsExtension does not show line numbers
